Split Flask app into modules
https://www.youtube.com/watch?v=44PvX0Yv368
