+++ backend/src/meetings/exceptions.py
@@ -1,0 +1,4 @@
+class ApiException(Exception):
+    def __init__(self, message, status_code):
+        super().__init__(message)
+        self.status_code = status_code
