+++ backend/src/meetings/run.py
@@ -1,0 +1,5 @@
+from src.meetings import app, db
+
+if __name__ == '__main__':
+    db.create_all()
+    app.run(debug=True, host='0.0.0.0', port=8000)
